+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-5471] Bulk update of quality implementations via csv
+- [TD-3541] Get data_structure_versions for `api/v2`
+
 ## [5.1.0] 2023-02-13
 
 ### Added
@@ -41,7 +48,7 @@
   their original `name` regardless of whether they have an alias or not. `alias`
   is indexed as an independent property if present.
 - [TD-5199] Soft delete for rule
-- [TD-5433] Ignore ReferenceData enrich in Implementation if it doesn't exists
+- [TD-5433] ignore ReferenceData enrich in Implementation if it doesn't exist
 - [TD-5432] Publish rule_result_created event using implementation_ref as
   resource_id instead of implementation_id.